Refactor drone to remove redundant 'drone' module
