+++ dev/tests/agent.rs
@@ -15,7 +15,7 @@ use dis_spawner::{
     types::{BackendId, ClusterName, DroneId},
 };
 use dis_spawner_drone::config::DockerConfig;
-use dis_spawner_drone::{database::DroneDatabase, drone::agent::AgentOptions, ip::IpSource};
+use dis_spawner_drone::{agent::AgentOptions, database::DroneDatabase, ip::IpSource};
 use integration_test::integration_test;
 use std::net::{IpAddr, Ipv4Addr};
 use std::time::Duration;
@@ -43,8 +43,7 @@ impl Agent {
             docker_options: DockerConfig::default(),
         };
 
-        let agent_guard =
-            expect_to_stay_alive(dis_spawner_drone::drone::agent::run_agent(agent_opts));
+        let agent_guard = expect_to_stay_alive(dis_spawner_drone::agent::run_agent(agent_opts));
 
         Ok(Agent {
             agent_guard,
