+++ drone/src/lib.rs
@@ -1,6 +1,9 @@
-pub mod acme;
+pub mod agent;
+pub mod cert;
+pub mod cli;
 pub mod config;
 pub mod database;
-pub mod drone;
 pub mod ip;
 pub mod keys;
+pub mod proxy;
+pub mod run;
