+++ drone/src/main.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use dis_spawner_drone::drone::run;
+use dis_spawner_drone::run::run;
 
 fn main() -> Result<()> {
     run()
